requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiobotocore==2.25.0
    # via s3fs
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.1
    # via
    #   aiobotocore
    #   fsspec
    #   s3fs
aioitertools==0.12.0
    # via aiobotocore
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via snowflake-ml-python
asn1crypto==1.5.1
    # via snowflake-connector-python
async-timeout==5.0.1
    # via aiohttp
attrs==25.4.0
    # via aiohttp
boto3==1.40.49
    # via snowflake-connector-python
botocore==1.40.49
    # via
    #   aiobotocore
    #   boto3
    #   s3transfer
    #   snowflake-connector-python
cachetools==5.5.2
    # via snowflake-ml-python
certifi==2025.10.5
    # via
    #   requests
    #   snowflake-connector-python
    #   snowflake-core
cffi==1.17.1
    # via
    #   cryptography
    #   snowflake-connector-python
charset-normalizer==3.4.4
    # via
    #   requests
    #   snowflake-connector-python
cloudpickle==3.1.1
    # via
    #   shap
    #   snowflake-ml-python
    #   snowflake-snowpark-python
cryptography==46.0.0
    # via
    #   pyopenssl
    #   snowflake-connector-python
    #   snowflake-ml-python
exceptiongroup==1.3.0
    # via anyio
filelock==3.20.0
    # via snowflake-connector-python
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2025.9.0
    # via
    #   s3fs
    #   snowflake-ml-python
idna==3.11
    # via
    #   anyio
    #   requests
    #   snowflake-connector-python
    #   yarl
importlib-resources==6.5.2
    # via snowflake-ml-python
jmespath==1.0.1
    # via
    #   aiobotocore
    #   boto3
    #   botocore
joblib==1.5.2
    # via scikit-learn
llvmlite==0.45.1
    # via numba
multidict==6.7.0
    # via
    #   aiobotocore
    #   aiohttp
    #   yarl
numba==0.62.1
    # via shap
numpy==2.2.6
    # via
    #   numba
    #   pandas
    #   scikit-learn
    #   scipy
    #   shap
    #   snowflake-ml-python
    #   xgboost
packaging==24.2
    # via
    #   shap
    #   snowflake-connector-python
    #   snowflake-ml-python
pandas==2.3.3
    # via
    #   shap
    #   snowflake-connector-python
    #   snowflake-ml-python
platformdirs==4.5.0
    # via
    #   snowflake-connector-python
    #   snowflake-ml-python
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.1
    # via snowflake-snowpark-python
pyarrow==18.1.0
    # via
    #   snowflake-connector-python
    #   snowflake-ml-python
pycparser==2.23
    # via cffi
pydantic==2.12.3
    # via
    #   snowflake-core
    #   snowflake-ml-python
pydantic-core==2.41.4
    # via pydantic
pyjwt==2.10.1
    # via
    #   snowflake-connector-python
    #   snowflake-ml-python
pyopenssl==25.3.0
    # via snowflake-connector-python
python-dateutil==2.9.0.post0
    # via
    #   aiobotocore
    #   botocore
    #   pandas
    #   snowflake-core
    #   snowflake-snowpark-python
pytimeparse==1.1.8
    # via snowflake-ml-python
pytz==2025.2
    # via
    #   pandas
    #   snowflake-connector-python
pyyaml==6.0.3
    # via
    #   snowflake-core
    #   snowflake-ml-python
    #   snowflake-snowpark-python
requests==2.32.5
    # via
    #   snowflake-connector-python
    #   snowflake-core
retrying==1.4.2
    # via snowflake-ml-python
s3fs==2025.9.0
    # via snowflake-ml-python
s3transfer==0.14.0
    # via boto3
scikit-learn==1.7.2
    # via
    #   shap
    #   snowflake-ml-python
scipy==1.15.3
    # via
    #   scikit-learn
    #   shap
    #   snowflake-ml-python
    #   xgboost
shap==0.49.1
    # via snowflake-ml-python
six==1.17.0
    # via python-dateutil
slicer==0.0.8
    # via shap
sniffio==1.3.1
    # via anyio
snowflake-connector-python[pandas]==3.18.0
    # via
    #   snowflake-core
    #   snowflake-ml-python
    #   snowflake-snowpark-python
snowflake-core==1.8.0
    # via snowflake-ml-python
snowflake-ml-python==1.17.0
    # via -r requirements.in
snowflake-snowpark-python==1.40.0
    # via
    #   -r requirements.in
    #   snowflake-ml-python
sortedcontainers==2.4.0
    # via snowflake-connector-python
sqlparse==0.5.3
    # via snowflake-ml-python
threadpoolctl==3.6.0
    # via scikit-learn
tomlkit==0.13.3
    # via snowflake-connector-python
tqdm==4.67.1
    # via
    #   shap
    #   snowflake-ml-python
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   cryptography
    #   exceptiongroup
    #   multidict
    #   pydantic
    #   pydantic-core
    #   pyopenssl
    #   shap
    #   snowflake-connector-python
    #   snowflake-ml-python
    #   snowflake-snowpark-python
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
tzlocal==5.3.1
    # via snowflake-snowpark-python
urllib3==2.5.0
    # via
    #   botocore
    #   requests
    #   snowflake-core
wheel==0.45.1
    # via snowflake-snowpark-python
wrapt==1.17.3
    # via aiobotocore
xgboost==3.1.1
    # via snowflake-ml-python
yarl==1.22.0
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
